This patch was done with the intention of allowing for more than one server to use the same bot. Perviously, because each server had its own emoji, the ID's for said emoji were different per server. That meant that when using the bot in seperate servers, it would print nonsence in the server that it wasn't set up with.
This patch uses the ID for the Server as a key for the emoji object. Now to add the bot to a second server, all you have to do it add the bot as a user, duplicate the object in staticValues.js, and fill it in with the second server's emoji IDs.

The purpose of this was to use the same bot in both my personal server, as well as a generic one for the public to use. Eventually, I'd love to add a DB component that allows for a single bot instance to be used in any server. This would require some kind of input from a user, to set the emoji tokens for their given server.